Wall Guns and Picking Up Ammo From Dropped Weapons - Pt 1
There was a little less development this week than in the past. I am starting to feel ready to move on to another project. But now that I am making progress on this again, I am feeling happy while working on it and am seeing the potential of this game again. :)