Wall Guns and Picking Up Ammo From Dropped Weapons - Changelog Note
Also, I forgot to mention previously that I changed how adding guns work. I now give a weapon a bullet type, which is a scene. This allows me to make weapons that shoot different kinds of bullets, whether it is a standard bullet with a different sprite or something completely new, like a grenade launcher!